created todo dto and endpoints, must be modified

+++ src/main/java/restmocker/backend/domain/TodoRepository.java
@@ -8,11 +8,11 @@ public interface TodoRepository {
 
   List<Todo> getTodos();
 
-  Todo getTodo(int id);
+  Todo getTodo(long id);
 
   Todo createTodo(Todo todo);
 
-  void updateTodo(int id, Todo todo);
+  void updateTodo(long id, Todo todo);
 
   void deleteTodos(List<Long> ids);
 }
